# ai-products-selection-frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/api/v1","description":"Generated server url"}],"paths":{"/weight-versions/{id}":{"put":{"tags":["weight-version-controller"],"operationId":"updateWeightVersion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeightVersionUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseWeightVersionDetail"}}}}}}},"/products/{id}":{"get":{"tags":["product-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductDetail"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductSaveResult"}}}}}},"delete":{"tags":["product-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/weight-versions":{"get":{"tags":["weight-version-controller"],"operationId":"listWeightVersions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListWeightVersionSummary"}}}}}},"post":{"tags":["weight-version-controller"],"operationId":"createWeightVersion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeightVersionUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseWeightVersionDetail"}}}}}}},"/products":{"get":{"tags":["product-controller"],"operationId":"listProducts","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"categoryId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"supplierId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"trackType","in":"query","required":false,"schema":{"type":"string","enum":["A","B"]}},{"name":"sourcingStatus","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","SOURCING","URGENT","PROMOTED","REJECTED"]}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["DRAFT","EVALUATING","WATCHING","ADOPTED","LISTED","REJECTED"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageResponseProductListItem"}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductSaveResult"}}}}}}},"/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTokenPair"}}}}}}},"/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResult"}}}}}}},"/risks/{id}/ignore":{"patch":{"tags":["risk-alert-controller"],"operationId":"ignore","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskIgnoreRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRiskAlertListItem"}}}}}}},"/risks/{id}/acknowledge":{"patch":{"tags":["risk-alert-controller"],"operationId":"acknowledge","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRiskAlertListItem"}}}}}}},"/products/{id}/status":{"patch":{"tags":["product-controller"],"operationId":"changeStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductStatusChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductDetail"}}}}}}},"/weight-versions/{id}/profiles":{"get":{"tags":["weight-version-controller"],"operationId":"profiles","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseWeightVersionDetail"}}}}}}},"/weight-versions/active":{"get":{"tags":["weight-version-controller"],"operationId":"active","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseWeightVersionDetail"}}}}}}},"/trends":{"get":{"tags":["trend-controller"],"operationId":"getTrends","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrendSignalProjection"}}}}}}}},"/trends/{keywordId}/detail":{"get":{"tags":["trend-controller"],"operationId":"getTrendDetail","parameters":[{"name":"keywordId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrendDetailResponse"}}}}}}},"/trends/{keywordId}/chart":{"get":{"tags":["trend-controller"],"operationId":"getTrendChart","parameters":[{"name":"keywordId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrendChartProjection"}}}}}}}},"/suppliers":{"get":{"tags":["supplier-controller"],"operationId":"listSuppliers","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListSupplierOption"}}}}}}},"/risks":{"get":{"tags":["risk-alert-controller"],"operationId":"listRisks","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["OPEN","ACKNOWLEDGED","IGNORED"]}},{"name":"severity","in":"query","required":false,"schema":{"type":"string","enum":["HIGH","MEDIUM","LOW"]}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"categoryId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageResponseRiskAlertListItem"}}}}}}},"/categories":{"get":{"tags":["category-controller"],"operationId":"listCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCategoryOption"}}}}}}},"/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthenticatedUser"}}}}}}}},"components":{"schemas":{"SceneWeightSet":{"type":"object","properties":{"sceneType":{"type":"string","enum":["VIRAL","FESTIVAL","REPLENISHMENT","SEASONAL"]},"weights":{"type":"object","additionalProperties":{"type":"number"}},"gradeAMin":{"type":"number"},"gradeBMin":{"type":"number"}},"required":["gradeAMin","gradeBMin","sceneType","weights"]},"WeightVersionUpsertRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":80,"minLength":0},"changeNote":{"type":"string","maxLength":512,"minLength":0},"effectiveFrom":{"type":"string","format":"date"},"scenes":{"type":"array","items":{"$ref":"#/components/schemas/SceneWeightSet"},"minItems":1}},"required":["name","scenes"]},"ApiError":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"fieldErrors":{"type":"array","items":{"$ref":"#/components/schemas/FieldError"}}}},"ApiResponseWeightVersionDetail":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/WeightVersionDetail"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"FieldError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"WeightVersionDetail":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"versionNo":{"type":"string"},"name":{"type":"string"},"status":{"type":"string","enum":["DRAFT","APPROVED","RETIRED"]},"effectiveFrom":{"type":"string","format":"date"},"current":{"type":"boolean"},"changeNote":{"type":"string"},"scenes":{"type":"array","items":{"$ref":"#/components/schemas/SceneWeightSet"}}}},"ProductRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"categoryId":{"type":"integer","format":"int64"},"supplierId":{"type":"integer","format":"int64"},"cost":{"type":"number"},"suggestedPrice":{"type":"number"},"moq":{"type":"integer","format":"int32"},"season":{"type":"string","enum":["ALL","SPRING","SUMMER","AUTUMN","WINTER","FESTIVAL"]},"trackType":{"type":"string","enum":["A","B"]},"sourcingStatus":{"type":"string","enum":["PENDING","SOURCING","URGENT","PROMOTED","REJECTED"]},"logisticsCondition":{"type":"string"},"shelfLifeDays":{"type":"integer","format":"int32"},"idealTempMin":{"type":"number"},"idealTempMax":{"type":"number"},"keywordIds":{"type":"array","items":{"type":"integer","format":"int64"}}},"required":["categoryId","name","trackType"]},"ApiResponseProductSaveResult":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/ProductSaveResult"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"ProductDetail":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"categoryId":{"type":"integer","format":"int64"},"categoryName":{"type":"string"},"supplierId":{"type":"integer","format":"int64"},"supplierName":{"type":"string"},"cost":{"type":"number"},"suggestedPrice":{"type":"number"},"marginRate":{"type":"number"},"moq":{"type":"integer","format":"int32"},"season":{"type":"string","enum":["ALL","SPRING","SUMMER","AUTUMN","WINTER","FESTIVAL"]},"status":{"type":"string","enum":["DRAFT","EVALUATING","WATCHING","ADOPTED","LISTED","REJECTED"]},"rejectReason":{"type":"string"},"listedAt":{"type":"string","format":"date"},"trackType":{"type":"string","enum":["A","B"]},"sourcingStatus":{"type":"string","enum":["PENDING","SOURCING","URGENT","PROMOTED","REJECTED"]},"logisticsCondition":{"type":"string"},"shelfLifeDays":{"type":"integer","format":"int32"},"idealTempMin":{"type":"number"},"idealTempMax":{"type":"number"},"lastScoringStatus":{"type":"string","enum":["SCORED","INSUFFICIENT_DATA"]},"lastScoringAttemptedAt":{"type":"string","format":"date-time"},"keywordIds":{"type":"array","items":{"type":"integer","format":"int64"}},"keywords":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ProductSaveResult":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductDetail"},"duplicateNameWarning":{"type":"boolean"}}},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"ApiResponseTokenPair":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/TokenPair"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"TokenPair":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"data":{},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ApiResponseLoginResult":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/LoginResult"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"AuthenticatedUser":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"displayName":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"LoginResult":{"type":"object","properties":{"tokens":{"$ref":"#/components/schemas/TokenPair"},"user":{"$ref":"#/components/schemas/AuthenticatedUser"}}},"RiskIgnoreRequest":{"type":"object","properties":{"reason":{"type":"string","maxLength":300,"minLength":0}},"required":["reason"]},"ApiResponseRiskAlertListItem":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/RiskAlertListItem"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"RiskAlertListItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"productName":{"type":"string"},"categoryId":{"type":"integer","format":"int64"},"categoryName":{"type":"string"},"riskType":{"type":"string"},"severity":{"type":"string","enum":["HIGH","MEDIUM","LOW"]},"triggerValue":{"type":"string"},"status":{"type":"string","enum":["OPEN","ACKNOWLEDGED","IGNORED"]},"ignoreReason":{"type":"string"},"detectedAt":{"type":"string","format":"date-time"},"handledAt":{"type":"string","format":"date-time"},"handledByName":{"type":"string"}}},"ProductStatusChangeRequest":{"type":"object","properties":{"status":{"type":"string","enum":["DRAFT","EVALUATING","WATCHING","ADOPTED","LISTED","REJECTED"]}},"required":["status"]},"ApiResponseProductDetail":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/ProductDetail"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseListWeightVersionSummary":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/WeightVersionSummary"}},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"WeightVersionSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"versionNo":{"type":"string"},"name":{"type":"string"},"status":{"type":"string","enum":["DRAFT","APPROVED","RETIRED"]},"effectiveFrom":{"type":"string","format":"date"},"current":{"type":"boolean"},"changeNote":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"approvedAt":{"type":"string","format":"date-time"}}},"TrendSignalProjection":{"type":"object","properties":{"aiSignal":{"type":"string"},"keywordId":{"type":"integer","format":"int64"},"keyword":{"type":"string"},"slope30d":{"type":"number"},"heatToday":{"type":"number"},"slope7d":{"type":"number"}}},"SourceDetail":{"type":"object","properties":{"sourceName":{"type":"string"},"percentile":{"type":"number","format":"double"},"actualWeight":{"type":"number","format":"double"},"status":{"type":"string"},"slope7d":{"type":"number","format":"double"},"slope30d":{"type":"number","format":"double"}}},"TrendDetailResponse":{"type":"object","properties":{"keyword":{"type":"string"},"heatToday":{"type":"number","format":"double"},"slope7d":{"type":"number","format":"double"},"slope30d":{"type":"number","format":"double"},"aiSignal":{"type":"string"},"sourceDetails":{"type":"array","items":{"$ref":"#/components/schemas/SourceDetail"}}}},"TrendChartProjection":{"type":"object","properties":{"heatScore":{"type":"number"},"date":{"type":"string","format":"date"}}},"ApiResponseListSupplierOption":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SupplierOption"}},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"SupplierOption":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"ApiResponsePageResponseRiskAlertListItem":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/PageResponseRiskAlertListItem"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"PageResponseRiskAlertListItem":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/RiskAlertListItem"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"ApiResponsePageResponseProductListItem":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/PageResponseProductListItem"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"PageResponseProductListItem":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductListItem"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"ProductListItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"categoryId":{"type":"integer","format":"int64"},"categoryName":{"type":"string"},"trackType":{"type":"string","enum":["A","B"]},"supplierId":{"type":"integer","format":"int64"},"supplierName":{"type":"string"},"cost":{"type":"number"},"suggestedPrice":{"type":"number"},"marginRate":{"type":"number"},"latestScore":{"type":"number"},"latestGrade":{"type":"string","enum":["A","B","C"]},"status":{"type":"string","enum":["DRAFT","EVALUATING","WATCHING","ADOPTED","LISTED","REJECTED"]},"sourcingStatus":{"type":"string","enum":["PENDING","SOURCING","URGENT","PROMOTED","REJECTED"]},"lastScoringStatus":{"type":"string","enum":["SCORED","INSUFFICIENT_DATA"]}}},"ApiResponseListCategoryOption":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryOption"}},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}},"CategoryOption":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"parentId":{"type":"integer","format":"int64"}}},"ApiResponseAuthenticatedUser":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/AuthenticatedUser"},"error":{"$ref":"#/components/schemas/ApiError"},"timestamp":{"type":"string","format":"date-time"}}}}}}