# ai-products-selection-frontend/src/app/core/api/model/apiResponseListWeightVersionSummary.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WeightVersionSummary } from './weightVersionSummary';
import { ApiError } from './apiError';


export interface ApiResponseListWeightVersionSummary { 
    success?: boolean;
    data?: Array<WeightVersionSummary>;
    error?: ApiError;
    timestamp?: string;
}