# ai-products-selection-frontend/src/app/core/api/model/apiResponseLoginResult.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LoginResult } from './loginResult';
import { ApiError } from './apiError';


export interface ApiResponseLoginResult { 
    success?: boolean;
    data?: LoginResult;
    error?: ApiError;
    timestamp?: string;
}