# ai-products-selection-frontend/src/app/core/api/model/apiResponseProductDetail.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductDetail } from './productDetail';
import { ApiError } from './apiError';


export interface ApiResponseProductDetail { 
    success?: boolean;
    data?: ProductDetail;
    error?: ApiError;
    timestamp?: string;
}