# ai-products-selection-frontend/src/app/core/api/model/apiResponseProductSaveResult.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { ProductSaveResult } from './productSaveResult';


export interface ApiResponseProductSaveResult { 
    success?: boolean;
    data?: ProductSaveResult;
    error?: ApiError;
    timestamp?: string;
}