# ai-products-selection-frontend/src/app/core/api/model/apiResponseRiskAlertListItem.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { RiskAlertListItem } from './riskAlertListItem';


export interface ApiResponseRiskAlertListItem { 
    success?: boolean;
    data?: RiskAlertListItem;
    error?: ApiError;
    timestamp?: string;
}