# ai-products-selection-frontend/src/app/core/api/model/apiResponseWeightVersionDetail.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WeightVersionDetail } from './weightVersionDetail';
import { ApiError } from './apiError';


export interface ApiResponseWeightVersionDetail { 
    success?: boolean;
    data?: WeightVersionDetail;
    error?: ApiError;
    timestamp?: string;
}