# ai-products-selection-frontend/src/app/core/api/model/fieldError.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FieldError { 
    field?: string;
    message?: string;
}