# ai-products-selection-frontend/src/app/core/api/model/productDetail.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductDetail { 
    id?: number;
    name?: string;
    categoryId?: number;
    categoryName?: string;
    supplierId?: number;
    supplierName?: string;
    cost?: number;
    suggestedPrice?: number;
    marginRate?: number;
    moq?: number;
    season?: ProductDetail.SeasonEnum;
    status?: ProductDetail.StatusEnum;
    rejectReason?: string;
    listedAt?: string;
    trackType?: ProductDetail.TrackTypeEnum;
    sourcingStatus?: ProductDetail.SourcingStatusEnum;
    logisticsCondition?: string;
    shelfLifeDays?: number;
    idealTempMin?: number;
    idealTempMax?: number;
    lastScoringStatus?: ProductDetail.LastScoringStatusEnum;
    lastScoringAttemptedAt?: string;
    keywordIds?: Array<number>;
    keywords?: Array<string>;
    createdAt?: string;
    updatedAt?: string;
}
export namespace ProductDetail {
    export type SeasonEnum = 'ALL' | 'SPRING' | 'SUMMER' | 'AUTUMN' | 'WINTER' | 'FESTIVAL';
    export const SeasonEnum = {
        All: 'ALL' as SeasonEnum,
        Spring: 'SPRING' as SeasonEnum,
        Summer: 'SUMMER' as SeasonEnum,
        Autumn: 'AUTUMN' as SeasonEnum,
        Winter: 'WINTER' as SeasonEnum,
        Festival: 'FESTIVAL' as SeasonEnum
    };
    export type StatusEnum = 'DRAFT' | 'EVALUATING' | 'WATCHING' | 'ADOPTED' | 'LISTED' | 'REJECTED';
    export const StatusEnum = {
        Draft: 'DRAFT' as StatusEnum,
        Evaluating: 'EVALUATING' as StatusEnum,
        Watching: 'WATCHING' as StatusEnum,
        Adopted: 'ADOPTED' as StatusEnum,
        Listed: 'LISTED' as StatusEnum,
        Rejected: 'REJECTED' as StatusEnum
    };
    export type TrackTypeEnum = 'A' | 'B';
    export const TrackTypeEnum = {
        A: 'A' as TrackTypeEnum,
        B: 'B' as TrackTypeEnum
    };
    export type SourcingStatusEnum = 'PENDING' | 'SOURCING' | 'URGENT' | 'PROMOTED' | 'REJECTED';
    export const SourcingStatusEnum = {
        Pending: 'PENDING' as SourcingStatusEnum,
        Sourcing: 'SOURCING' as SourcingStatusEnum,
        Urgent: 'URGENT' as SourcingStatusEnum,
        Promoted: 'PROMOTED' as SourcingStatusEnum,
        Rejected: 'REJECTED' as SourcingStatusEnum
    };
    export type LastScoringStatusEnum = 'SCORED' | 'INSUFFICIENT_DATA';
    export const LastScoringStatusEnum = {
        Scored: 'SCORED' as LastScoringStatusEnum,
        InsufficientData: 'INSUFFICIENT_DATA' as LastScoringStatusEnum
    };
}