# ai-products-selection-frontend/src/app/core/api/model/productListItem.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductListItem { 
    id?: number;
    name?: string;
    categoryId?: number;
    categoryName?: string;
    trackType?: ProductListItem.TrackTypeEnum;
    supplierId?: number;
    supplierName?: string;
    cost?: number;
    suggestedPrice?: number;
    marginRate?: number;
    latestScore?: number;
    latestGrade?: ProductListItem.LatestGradeEnum;
    status?: ProductListItem.StatusEnum;
    sourcingStatus?: ProductListItem.SourcingStatusEnum;
    lastScoringStatus?: ProductListItem.LastScoringStatusEnum;
}
export namespace ProductListItem {
    export type TrackTypeEnum = 'A' | 'B';
    export const TrackTypeEnum = {
        A: 'A' as TrackTypeEnum,
        B: 'B' as TrackTypeEnum
    };
    export type LatestGradeEnum = 'A' | 'B' | 'C';
    export const LatestGradeEnum = {
        A: 'A' as LatestGradeEnum,
        B: 'B' as LatestGradeEnum,
        C: 'C' as LatestGradeEnum
    };
    export type StatusEnum = 'DRAFT' | 'EVALUATING' | 'WATCHING' | 'ADOPTED' | 'LISTED' | 'REJECTED';
    export const StatusEnum = {
        Draft: 'DRAFT' as StatusEnum,
        Evaluating: 'EVALUATING' as StatusEnum,
        Watching: 'WATCHING' as StatusEnum,
        Adopted: 'ADOPTED' as StatusEnum,
        Listed: 'LISTED' as StatusEnum,
        Rejected: 'REJECTED' as StatusEnum
    };
    export type SourcingStatusEnum = 'PENDING' | 'SOURCING' | 'URGENT' | 'PROMOTED' | 'REJECTED';
    export const SourcingStatusEnum = {
        Pending: 'PENDING' as SourcingStatusEnum,
        Sourcing: 'SOURCING' as SourcingStatusEnum,
        Urgent: 'URGENT' as SourcingStatusEnum,
        Promoted: 'PROMOTED' as SourcingStatusEnum,
        Rejected: 'REJECTED' as SourcingStatusEnum
    };
    export type LastScoringStatusEnum = 'SCORED' | 'INSUFFICIENT_DATA';
    export const LastScoringStatusEnum = {
        Scored: 'SCORED' as LastScoringStatusEnum,
        InsufficientData: 'INSUFFICIENT_DATA' as LastScoringStatusEnum
    };
}