# ai-products-selection-frontend/src/app/core/api/model/tokenPair.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenPair { 
    accessToken?: string;
    refreshToken?: string;
}