# ai-products-selection-frontend/src/app/core/api/model/weightVersionSummary.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WeightVersionSummary { 
    id?: number;
    versionNo?: string;
    name?: string;
    status?: WeightVersionSummary.StatusEnum;
    effectiveFrom?: string;
    current?: boolean;
    changeNote?: string;
    createdAt?: string;
    approvedAt?: string;
}
export namespace WeightVersionSummary {
    export type StatusEnum = 'DRAFT' | 'APPROVED' | 'RETIRED';
    export const StatusEnum = {
        Draft: 'DRAFT' as StatusEnum,
        Approved: 'APPROVED' as StatusEnum,
        Retired: 'RETIRED' as StatusEnum
    };
}